WIP: adv drive & driveapp

+++ src/support/filecache.js
@@ -3,6 +3,7 @@
  * so we'll cache that over here
  */
 
+
 const _performance = {
   hits: 0,
   misses: 0,
@@ -13,6 +14,7 @@ export const getPerformance = () => _performance
 
 const CACHE_ENABLED = true
 const fileCache = new Map()
+
 const _getFromCache = (id) => {
   if (CACHE_ENABLED) {
     const has = fileCache.has (id)
@@ -25,25 +27,55 @@ const _getFromCache = (id) => {
   }
   return null
 }
-export const setInFileCache = (id, file) => {
-  if (CACHE_ENABLED) {
+
+export const improveFileCache = (id, file) => {
+
+  if (CACHE_ENABLED && id) {
+    const cachedFile = fileCache.get (id) || {}
+    const improved = {...cachedFile, ...file}
+    file = setInFileCache (id, improved)
+  }
+  return file
+}
+
+export const setInFileCache = ( id, file) => {
+
+  if (CACHE_ENABLED && id) {
     if (fileCache.has(id) && file === null || typeof file === typeof undefined) {
       fileCache.delete(id)
     } else {
-      fileCache.set(id, file)
+      // fake the parents if not given because that'll be the root
+      fileCache.set(id, {parents:[], ...file})
     }
   }
   return file
 }
 
-export const getFromFileCache = (id, fields = '') => {
-  fields = fields.split (",")
+export const getFromFileCache = (id, fields = []) => {
+
+  // get the thing from cache
   const cachedFile = _getFromCache(id)
-  if (cachedFile && (!fields || fields.every(f => Reflect.has(cachedFile, f)))) {
-    _performance.fieldHits++
-    return cachedFile
-  } else {
-    if (cachedFile) _performance.fieldMisses++
-    return null
+
+  if (cachedFile) {
+    // if we already have all the fields needed
+    if (fields.every(f => Reflect.has(cachedFile, f))) {
+      _performance.fieldHits++
+      return {
+        cachedFile,
+        good: true
+      }
+    }
+    _performance.fieldMisses++
+    return {
+      cachedFile,
+      good:false
+    }
   }
+  // missed or didnt have enough fields
+  return {
+    cachedFile,
+    good: false
+  }
+
+
 }
